requirements.txt
openai
RPi.GPIO
requests
Pillow
pyserial
Unidecode
os